requirements.txt
pyserial
PyQT5
